requirements.txt
asgiref==3.2.7
certifi==2019.9.11
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-stdimage==5.1.1
gunicorn==20.0.4
Pillow==7.1.2
pipenv==2018.11.26
progressbar2==3.51.3
psycopg2-binary==2.8.5
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
